+++ R/metadata.R
@@ -12,7 +12,7 @@ num_slices <- function(img_data) {
 
 #' Get the dimensions of an image
 #' @param img_data Image data returned by e.g. \code{\link{read_dicom}} or \code{\link{read_nifti1}}
-#' @return Image dimensions
+#' @return Image dimensions, or NA if image has no slices
 #' @examples
 #' img_dimensions(radtools::sample_dicom_img)
 #' img_dimensions(radtools::sample_nifti_img)