+++ NEWS.md
@@ -1,3 +1,11 @@
+# radtools 1.0.2
+
+Support DICOM objects with no image data
+
+# radtools 1.0.1
+
+* Add Zenodo archiving
+
 # radtools 1.0.0
 
 * Initial release.